consensus/account_list: refactor into top-level analyzer
