+++ tests/e2e_regression/eden/e2e_config_2.yml
@@ -13,6 +13,7 @@ analysis:
     caching_proxies:
       - { target: "https://sourcify.dev/server", host_addr: "localhost:9191" }
   analyzers:
+    consensus_accounts_list:       { interval: 5s, stop_if_queue_empty_for: 1s }
     evm_tokens_emerald:            { stop_if_queue_empty_for: 1s }
     evm_nfts_emerald:              { stop_if_queue_empty_for: 1s }
     evm_token_balances_emerald:    { stop_if_queue_empty_for: 1s }
